[BUGFIX] selicon_field does not create thumbnails
A select field with renderType selectSingle can point to a
foreign_table. If on the foreign_table TCA the ctrl keys
selicon_field and selicon_field_path are set, a single foreign row
can have an "icon" field that is shown to represent this row.

The patch fixes a bug where selicon_field of the own table
instead of the foreign table was used.

Resolves: #75577
Releases: master, 7.6
Change-Id: I50bac28018b17a61a334aac7d241bcdd96663656
Reviewed-on: https://review.typo3.org/47887
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>