[TASK] Doctrine: Add missing PHP7 type hints to QueryContext
The PHP7 return type declaration for getIgnoredEnableFields() as well as
the input type hint for setIgnoredEnableFields() have been added.

Additionally the development notes at the top of the class have been
removed as all information from those notes was included in the
appropriate phpDoc headers of the methods.

Releases: master
Resolves: #75574
Change-Id: I9765be595e1e3f2eb16fc5f3ba51fdab3681fec2
Reviewed-on: https://review.typo3.org/47649
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>