[!!!][TASK] Always store <p> tags in DB from RTE
Twelve years ago, when the RTE was not working in
every browser, TYPO3 always stored as little
html code in the database as possible.

Now it's time to really store HTML in the database
so the custom transformations from RTE to DB
(and vice versa) and from DB to the frontend
are getting less.

This introduces less overhead and also makes
it easier to work with the content that
is stored in the database.

An upgrade wizard (coming separately) will ensure
that all existing fields will be migrated.

Resolves: #75708
Releases: master
Change-Id: I47ffb394129e6a6e5a2916616931c050f7b9b064
Reviewed-on: https://review.typo3.org/47721
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Michiel Roos <michiel@maxserv.com>
Tested-by: Michiel Roos <michiel@maxserv.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>