Revert "[BUGFIX] config.locale_all should format floating point values"
PHP 7 still has the problem converting a float to a string is
locale-aware (https://bugs.php.net/bug.php?id=53711). This means a float
might be returned with a comma as decimal point. This is not intended
and breaks value conversions so this patch restores the old behavior.

This reverts commit 7cbef93a8698347bc55e364fb7ffa3e796b733ec.

Resolves: #75780
Resolves: #75825
Releases: master, 7.6
Change-Id: I83a443617e50a86b0297c49887e94f226104c910
Reviewed-on: https://review.typo3.org/47828
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Tested-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>