[BUGFIX] Clear cache inconsistency in toolbar
With the simplification of the clear cache items in
toolbar the clearCacheSystem TYPO3_CONF_VARS was
rendered obsolete.
The patch removes last occurences in toolbar and
simplifies the visibility restrictions again: The
clear all toolbar item is now shown if user is admin
and the TSconfig option is not explictly disabling
the option for an admin, or a non-admin has the
TSconfig options.clearCache.all set to 1.

Change-Id: I5608baf2d4fe0c21dcc38f09fc17e550fcbfe8ff
Resolves: #75844
Releases: master
Reviewed-on: https://review.typo3.org/47835
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>