Revert "[BUGFIX] config.locale_all should format floating point values"
PHP 7 still has the problem converting a float to a string is
locale-aware (https://bugs.php.net/bug.php?id=53711). This means a float
might be returned with a comma as decimal point. This is not intended
and breaks value conversions so this patch restores the old behavior.

This reverts commit 7cbef93a8698347bc55e364fb7ffa3e796b733ec.

Resolves: #75780
Resolves: #75825
Releases: master, 7.6
Change-Id: I83a443617e50a86b0297c49887e94f226104c910
Reviewed-on: https://review.typo3.org/47841
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>