[BUGFIX] Respect page TSConfig pid overrides for new inline children
TCAdefaults.<table>.pid = <page id> in page TSConfig can be used for
new inline children to define a pid new records should be located at,
even if the parent record is on a different page. This can be useful
when having special storage folders on a per-table-basis.

The patch fixes this feature that broke during FormEngine refactoring.

Change-Id: I8e60155612397e72f1b11a068617f88e3793384b
Resolves: #70780
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/47905
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>