[BUGFIX] Fix adding fields to an existing table in a specific position
When trying to add new fields at a specific position to an existing
table (e.g. "after:title") and the field referenced in the position is
part of a palette the regular expression to insert the new fields did
not work.

This happened because a regular expression part was inserted in the
pallete string to replace and being quoted afterwards. That caused the
preg_replace function to treat the regular expression as a string as it
has been escaped.

Resolves: #75946
Releases: master,7.6
Change-Id: Ib0dbcb56e530e1671470210218bd1c1205b50946
Reviewed-on: https://review.typo3.org/47935
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>