[BUGFIX] Exception editing inline mm with deleted child child
Have an inline m:m record and delete one child child that has an
intermediate record pointing to it. Opening the parent throws
a DatabaseRecordException.
The patch extends this exception to add tableName and uid, then
catches the exception in the inline data provider, creates a
nice error message as flash message and continues displaying record.

Change-Id: I1792716b4e5454b11499cb2ba684bac403b3f13d
Resolves: #71719
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/47969
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>