[BUGFIX] RTE: Determine correct context menu position
The context menu of the RTE is now correctly positioned with this patch.
The position is now based on the cursor's position. The `targetPosition` is
removed now as it has strange side-effects if the context menu is opened
at formatted text.

The method `constrainScroll` is overridden to circumvent a bug in ExtJS
where the menu is repositioned and the height is changed to show
as much items as possible, even if the ``maxHeight`` setting is set.

Resolves: #74332
Releases: master, 7.6
Change-Id: I9e2e42c98ca578101945f65f7c50e32a615806b7
Reviewed-on: https://review.typo3.org/47977
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>