[BUGFIX] Output HTML code only if requested
When adding visibility information in DebuggerUtility
it was not respected that this utility can also be used to output in cli.

Add the missing condition and output HTML only if requested.

Additionally when the property has nested objects,
the visibility info must be right after the property not after
every nested object.

Also prettify object type output on command line on the go.

Resolves: #76301
Related: #76008
Releases: master
Change-Id: If82192bf9d1fb0ca1dc843242167d01b1a915f6e
Reviewed-on: https://review.typo3.org/48287
Reviewed-by: Elmar Hinz <t3elmar@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Elmar Hinz <t3elmar@gmail.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>