[BUGFIX] Fix output of DebugUtility
This change prettifies the output of the debug output
methods, by using the Extbase DebuggerUtility for that purpose.

The benefits are consistent, well readable, properly encoded output
for every case the class provides.

In the long run, this class should be replaced with
a more object oriented approach.

Resolves: #76302
Releases: master, 7.6
Change-Id: Iacbb48701f6c98139bd7db86795952a123e076da
Reviewed-on: https://review.typo3.org/48398
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>