[TASK] Obsolete unserialize(serialize())
There is a funny comment for this call within TypoScriptParser.
Reading the code, there seems to be no part that works with
references on this variable in setVal().

The patch now removes this odd construct. In the
unlikely case some issue still pops up later, we would
at least have a clear way to reproduce then ...

Change-Id: I70f5f8915386a7a4b5038d66611b4ec359f5c1fa
Resolves: #76323
Releases: master
Reviewed-on: https://review.typo3.org/48327
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>