[TASK] Install Tool: Store numbers and bool values properly
The install tool ("All Configuration") does not recognize boolean
or integer values, even though they are set in the comments section
of the DefaultConfiguration.php file.

This is changed into bool and even integer numbers to be cast correctly
when using All Configuration and storing the data into LocalConfiguration.php.

Resolves: #76336
Releases: master
Change-Id: I919d2390fd9e07438f77cfd7b76b2d19b6b08000
Reviewed-on: https://review.typo3.org/48345
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>