[TASK] Make base test classes always available
To ease the usage of the core with extensions,
we move the base test classes into a different place
to be able to make them available when typo3/cms is pulled in
as dependency, e.g. for testing extensions or in projects.

This is a non breaking change, as the classes will be available
in any case after that.

By doing so, we can nicely cleanup the .gitattributes file

Resolves: #76425
Releases: master, 7.6
Change-Id: I639f08a4e3e5dd73aa47570d326b72c64452acaf
Reviewed-on: https://review.typo3.org/49159
Reviewed-by: Peter Foerger <pfoerger@mac.com>
Tested-by: Peter Foerger <pfoerger@mac.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Timo Hund <timo.hund@yahoo.com>
Tested-by: Timo Hund <timo.hund@yahoo.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>