[BUGFIX] Fix some sql errors running functional tests
When executing the functional tests, the setUp() method
uses the extbase ObjectManager to populate the database schema.
The ObjectManager however relies on ready database schema
in the object reflection. This chicken-egg situation triggers
a lot of sql errors like "table cf_extbase_object does not exist".
For now, the solution is to switch to the NullBackend for this
cache in functional tests LocalConfiguration, since using
the object manager can not be dropped easily.
travis-ci does not show these failing queries due to the
parellelization, this part, together with other failing
queries should be fixed with further patches.

Change-Id: Ic41fbcf44d8daa368985a7c6c389ffd0d61e3450
Resolves: #76573
Releases: master
Reviewed-on: https://review.typo3.org/48546
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>