[BUGFIX] Handle missing option values in CHECKBOXGROUP and RADIOGROUP
Options are now checked for missing values and empty values are
inserted instead.

In rebuild process RADIOGROUP and CHECKBOXGROUP for every listed option
a single RADIO/CHECKBOX element is created by its xtype-configuration.
These elements rely on a passed value attribute. To prevent errors,
triggered by missing value attributes, a check for existence of the
attribute was introduced. In case of missing value, now an empty value
is used.
This change only influences the display process of the wizard.
Generation of Form-Configuration is not changed.

Resolves: #76019
Releases: master, 7.6
Change-Id: I3f459be2bec92c9eac20fd45a7453dbe6e28f758
Reviewed-on: https://review.typo3.org/48721
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>