[TASK] Smarter functional test parallelization
Currently travis creates one phpunit process per functional test
file and runs them through parallel. As a disadvantage
the list of functionals in travis is long and hard to read.

The new script finds all test files and weights them depending
on the number of tests in it. It then creates a number of
configuration files with a roughly even share of tests in
each run.

In effect, the travis output is now much better readable and
this script can be used in bamboo for execution on different
agents at the same time, too.

Change-Id: If70bba4bd282d486c19d5796494a5940602ab466
Resolves: #76759
Releases: master
Reviewed-on: https://review.typo3.org/48656
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>