[TASK] RTEHtmlArea: Remove inclusion of ext_emconf.php
In every request, EXT:rtetmlarea's ext_localconf.php is including
the ext_emconf.php file of the extension to fetch the conflicts (for
EXT:reports) and the rtehtmlarea version, to display in the AboutEditor
plugin of RTEHtmlArea.

The patch changes the version display of the extension to use
the TYPO3 version (which is equivalent to the extension version ATM)
and puts the conflicts check inside the Status Report directly
so it is only used at the places where needed, thus, speeding up every
page request of TYPO3 a bit more.

Resolves: #76880
Releases: master
Change-Id: Ica9fa8856a3057d2cdc3727b060a7df95648209e
Reviewed-on: https://review.typo3.org/48737
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>