[BUGFIX] Adodb: PHP 7.1 compatibility
unset($this) throws a fatal if linting with PHP 7.1.
According to php.net/unset the construct isn't allowed since
PHP 5. Now the linter stumbles upon that, too.
Solution is to drop these calls, the according files and
classes from adodb are not used by our dbal integration anyway.
A pull request to upstream is pending for this, too, but we
never know when it will be merged. Should be safe to do that
locally in our adapted version.

Change-Id: Ieeffee67c82dc9c7c9fef519ffea59685e48b37b
Resolves: #76947
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/48811
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>