[!!!][TASK] Remove legacy soft references functionality
The soft references based on TSconfig, TStemplate and images within
fileadmin/ are not accurate and thus useless since it's not common
anymore to reference TSconfig or TStemplate code inline inside the
DB fields, and also the data inside the softrefparser hasn't been used
anymore.

In order to refactor the SoftReference Index the old legacy code is
removed first without substitution.

The following softref keys are not evaluated anymore:
* TSconfig
* TStemplate

and can be removed from TCA.

The "image" soft reference key is not evaluating files within
fileadmin/ anymore (that are not linked as files).

The "typolink" and "typolink_tag" soft reference parsers don't evalute
links to plain files within fileadmin/ anymore.

The public property "SoftReferenceIndex::$fileAdminDir" is removed.

The following PHP methods are removed without substitution:
* SoftReferenceIndex::findRef_TStemplate()
* SoftReferenceIndex::findRef_TSconfig()
* SoftReferenceIndex::fileadminReferences()

Resolves: #77156
Releases: master
Change-Id: I0bc791a051fda06715175c2f526dddb90b8e4391
Reviewed-on: https://review.typo3.org/48933
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>