[BUGFIX] Fixes the suggest wizard item click by only returning one value
The bug only occurs on cases where the multiple true is defined,
and the user clicks on the element found by the suggest wizard.

This is caused because of a duplicate item insert. The onclick should
only prevent the default action, and not insert/return the item into the
list. 



Resolves: #76923
Releases: master, 7.6
Change-Id: Iea53ef266e8d55331f5cd8367a5146582b7fe684
Reviewed-on: https://review.typo3.org/49223
Reviewed-on: https://review.typo3.org/49862
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>