[BUGFIX] Adding back access to parentMenuArr and menuitem in subMenu
Adding the possibility to access the menuArr and direct menuItem from
the parent menu in subMenu instances to provide the ability to check
for parent item information or parent sibling item informations.

Providing two getters and a combined setter, called in the subMenu
method to readd this data access instead of simple using the old
unsafe and undeclared property set.

Leaving the parentMenuArr member public for now, even it is not
nice.

Resolves: #77296
Releases: master, 7.6
Change-Id: I3cfce3c67e840e6e97fa9c55ffaf997a509ee0e8
Reviewed-on: https://review.typo3.org/50376
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>