[TASK] Doctrine: Migrate a simple upgrade wizard
Migrate a simple install tool upgrade wizard that performs
an update query on be_users.
The upgrade wizard have a feature to display performed queries.
Usually, with doctrine, values would be prepared statement
values and getSQL() would display a placeholder only.
This is worked around here giving set() the third parameter
boolean false and quoting manually.

For the wizards itself, there are two nearly identical ones,
the patch merges both and improves the detection if an
execution is needed an no longer relies on "was executed once"
flag.

Change-Id: Ifcfcf75731076ed9f4b20981ea2be71ec672f47f
Resolves: #77446
Releases: master
Reviewed-on: https://review.typo3.org/49426
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>