[!!!][TASK] Drop extbase query cache
The database is having a query cache to exactly take care of that,
no need to cache a preparsed query (done in PHP) to be stored in
a different DB table by default to optimize speed.

Any interface related method and properties throw a deprecation
notice but will have no effect anymore.

The interface method has been removed, being compatible with
all existing older versions of the interface.

Resolves: #77460
Releases: master
Change-Id: I11e9d6c367a32fad2fa77ecd85172fc242ca1fe9
Reviewed-on: https://review.typo3.org/49409
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>