[BUGFIX] Continue searching after empty result
Change the default behaviour of jQuery autocomplete to continue
searching for characters even if the search already returned nothing.

The standard behaviour of autocomplete is to stop searching after
no results were found.
However, this made it impossible to use commands.
Searching for '#content:12345' did not work when '#con' could not
be found. Search was not executed for subsequent characters.

The parameter 'preventBadQueries: false' is used to force
searching for every character entered.

Resolves: #77566
Releases: master, 7.6
Change-Id: I423df51ae34ae258c42a418040d24cd91886872a
Reviewed-on: https://review.typo3.org/49562
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>