[FEATURE] Doctrine: Implement SchemaMigrationService
Implement a SQL schema migration service based on an actual
parser for CREATE TABLE statements that are mapped to Doctrine
Table objects. This enables the use of the Doctrine DBAL
SchemaManager for all schema modifications.

The new Schema migration service is fully aware of multiple
database connections and normalizes MySQL specific data types
to standard compliant types. This mostly affects the TINYINT
data type which gets converted to a SMALLINT.

Resolves: #77643
Resolves: #77369
Resolves: #76508
Resolves: #76641
Resolves: #75205
Resolves: #71645
Resolves: #44991
Releases: master
Change-Id: Ic56941c2ae9717836d89bce74261d11424da340b
Reviewed-on: https://review.typo3.org/49593
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>