[!!!][TASK] Migrate Extbase Query Parser to Doctrine DBAL
The Extbase Typo3DbQueryParser which builds the SQL
statement for most of Extbase's database calls is completely
rewritten to use Doctrine DBAL QueryBuilder.

The QueryParser now returns a properly filled QueryBuilder
object instead of an array with SQL parts.

The Typo3DbBackend of Extbase is using the QueryBuilder
under the hood then.

Resolves: #77379
Releases: master
Change-Id: I5936e639a9241a7d41ac60703efed83bda73f5f7
Reviewed-on: https://review.typo3.org/49584
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>