[TASK] Limit request to get logout information
Currently a request to get information about the login state is
triggered every minute. To prevent traffic the request can be triggered
only when the session is about to expire. This patch adds a function
to be able to pass the interval time to the LoginRefresh refresh and
sets it properly after initialization.

Resolves: #76364
Releases: master, 7.6
Change-Id: I9bfeb578973f1a5bc3a0f662ac1f6dcce82e7d37
Reviewed-on: https://review.typo3.org/48393
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>