[BUGFIX] Only check if a record lives in a workspace if extension is installed
The reference index checks if a record type is enabled for workspaces and
if a record exists in the currently selected workspace. However, this is
not required if EXT:workspaces is not installed at all as a lot of useless
queries could be run.

It's now checked whether the extension is installed before doing further
workspace processing of the record.

Resolves: #77676
Related: #64253
Releases: master, 7.6
Change-Id: I94aab84fce72e8b852783820b34065ab1e051ea5
Reviewed-on: https://review.typo3.org/49628
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>