[BUGFIX] Exclude uid '0' from any editing action in Info > Pagetree Overview
The module Info > Pagetree Overview offers the possibility to edit the
root page (uid 0) which is impossible as this is no real page.

The edit button for the row in the table representing the root page is
removed. The mass edit buttons in the table header don't allow to edit
that page anymore, too.

Resolves: #77678
Releases: master, 7.6
Change-Id: I7083c83372ad619ea8e8ffd2f53822bcd939d768
Reviewed-on: https://review.typo3.org/49659
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>