[CLEANUP] Change Enumeration::cast return doc to self
The Enumeration implementation has a cast() method that
does always return an instance of the enumeration
implementation itself. As the Enumeration is an abstract
type the cast() method return doc should be `self` instead
of `Enumeration`.
That does fix problems with code completion in IDEs as the
IDE does now know what exact implementation is returned.

Resolves: #77743
Releases: master, 7.6
Change-Id: Ib0f62f79e0856fe62849472a0a9c3657a048e028
Reviewed-on: https://review.typo3.org/49717
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Tested-by: Bamboo TYPO3com <info@typo3.com>