[BUGFIX] Do not throw exception for inaccessible folders
In the LocalDriver the PHP RecursiveIteratorIterator implementation is
used to recursively iterate over a directory.
If there is a resource child with limited access an exception is thrown
during the initialization of the iterator.
That means you are even not able to iterate over the other folders.

This patch adds a flag that prevents the iterator to throw the exception.

Resolves: #71328
Releases: master, 7.6
Change-Id: I3373f70bbcd5cdb38e5f8998a1e4c0466f6d5260
Reviewed-on: https://review.typo3.org/50128
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>