[BUGFIX] IRRE file upload button not shown to BE user
If the logged in BE user does not have the "addFolder"
file operation permission, the button for uploading
files in IRRE forms is not shown. This commit fixes
this by checking for the "addFile" permission instead.

When fetching the default upload folder, TYPO3 already
checks whether the user is allowed to write to that
particular folder.

Releases: master,7.6
Resolves: #77841
Change-Id: Ic76d5fb36fd39b9fbcc554b9f03517c3cb1f301a
Reviewed-on: https://review.typo3.org/49870
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>