[BUGFIX] Prevent duplicate pastes in rtehtmlarea
if pasteFormat or pasteStructur is activated in user settings,
pastings in rtehtmlarea will be inserted twice. Once as plaintext
and a second time as html.
Instead of appending all different text types of ClipBoard
we now override the complete clipboardText, if it is not
of the expected type.
To prevent pasting header data from OpenOffice/LibreOffice/Word
we have added a default value for removeTagsAndContents
which is configurable now.

Resolves: #67661
Releases: master,7.6
Change-Id: I62418ed2d1ce72bb1fb2839e23eaeaa76f42a4c8
Reviewed-on: https://review.typo3.org/49871
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>