[!!!][SECURITY] Mitigate potential cache flooding
Bind cHash to the page id it was generated for, to avoid
an attacker to be able to call multiple pages with the same
cHash arguments and thus create unnecessary cache entries.

We now add the id argument to the cHash calculation, but only
if there are other arguments in the URI which would require a cHash.
This avoids multiple cache entries for one page
(one with and one without cHash).

We ignore other core parameters like "type" and "MP", because the possibility
to create unnecessary cache entries by manipulating these is very limited
and thus an attack not feasible.

Adapted tests to show our new expectations for cHash calculations.

The new behavior is default for new installations, but not for on for existing
installations, as an update would break the site with a high probability.

By adding the configuration option, we'll give users the chance to
pull the trigger once everything is prepared, but still get other
security issues fixed with the release.

Resolves: #76462
Releases: master, 8.3, 7.6, 6.2
Security-Commit: 5b0a8866200bca3bdf021fd0979e8007b4117218
Security-Bulletins: TYPO3-CORE-SA-2016-020, 021
Change-Id: Id1c1c82e2a88481faede5898d055794fb4468c5b
Reviewed-on: https://review.typo3.org/49923
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>