[BUGFIX] Always throw exception if ext_emconf.php is missing
The ext_emconf.php is still required for TYPO3 extensions.
However a piece of PackageManager code was still graceful
if instead at least a composer.json was present.

Then however it failed, when in composer.json the require section
was filled with composer packages which are unknown to TYPO3.

Because of this, we need to throw an exception in any case,
if ext_emconf.php is not present.

Resolves: #78054
Releases: 7.6, master
Change-Id: I08b9931dac50384e08fe89558288fa8653722a78
Reviewed-on: https://review.typo3.org/50050
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>