[BUGFIX] Exception with cHashIncludePageId but no id in the URL
If the 'id' parameter is not provided in the URL but the
FE|cHashIncludePageId is set to true, TYPO3 will throw an exception. If
id is missing in $_GET, TYPO3 should use the determined id for the
check because in such case the id is assumed to be the uid of the home
page of the current site.

Resolves: #78021
Releases: master, 7.6, 6.2
Change-Id: Ib13d7aac9a75b2be54c7fddf6896af2f0a52e18e
Reviewed-on: https://review.typo3.org/50030
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Allacher <andreas.allacher@gmx.at>
Tested-by: Andreas Allacher <andreas.allacher@gmx.at>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Helmut Hummel <typo3@helhum.io>