[BUGFIX] Incorrect cHash generation may cause 404 on any page
If an empty parameter appears in the query string passed to
CacheHashCalculator::getRelevantParameters(), the code will
calculate cHash even if there are no other parameters. But
upon going to the generated link, there will be a 404 error
due to failed cHash check. Parameters with trimmed empty
names do not appear in $_GET in PHP and cHash cannot be
validated. The code should ignore the parameter with
trimmed empty name.

Resolves: #78102
Releases: master, 6.2, 7.6
Change-Id: I3c43d62ae9c31bd41674db62090b6b4a520f8b68
Reviewed-on: https://review.typo3.org/50070
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>