[FOLLOWUP][BUGFIX] Fix evaluation of rootLevel configuration in NewRecordController
If there is no configuration for the rootLevel in a tables' TCA, the
NewRecordController silently assumes -1. This means a table can be
inserted on a page AND on root level. This behaviour is wrong because in
TCA documentation per default rootLevel is 0 which means, this table can
only be used on pages.
This patch sets the correct default value for rootLevel if not set.

Resolves: #78170
Releases: master, 7.6
Change-Id: I3a34eb5abbd95b7ee60b8b676c86abaacd0e2a18
Reviewed-on: https://review.typo3.org/50102
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>