[TASK] Remove acceptance test setup from travis
The acceptance tests are very hard to stabilize on travis
and still flaky. They are disabled for quite a while for
that reason already.
The setup is now removed from travis to not clutter the
core and the config with not needed stuff.
In the unlikely case someone works on that again, they
could still be added again.

Change-Id: I95575b84e126216ec0483d2b060cafcc023e0317
Resolves: #78280
Releases: master
Reviewed-on: https://review.typo3.org/50211
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>