[BUGFIX] Remove dependency to RsaEncryptionModule in LoginRefresh
This patch removes the hard dependency to RsaEncryptionModule and adds
a check if the module can be loaded.

It on the go adds a hardening to the RsaEncryptionModule to avoid
registration of form event handlers twice, which is now needed to cover
all potential loading orders of RsaEncryptionModule and LoginRefresh. 

Resolves: #78299
Related: #75911
Releases: master, 7.6
Change-Id: Ie03f1c7bc34e48f03213dec70c62d8ccc339ab31
Reviewed-on: https://review.typo3.org/50239
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>