[BUGFIX] Implement CMS setRenderingContext for AbstractConditionViewHelper
This fixes an issue where the CMS specific overridden
setRenderingContext method on CMS AbstractConditionViewHelper is
not present because the class does not subclass the CMS
AbstractViewHelper (due to coupling concerns). Subsequently the
ControllerContext is not assigned in $this->controllerContext.

This inconsistency is fixed by adding the method just like it
appears in the AbstractViewHelper class, along with declaring
the property containing the ControllerContext.

Change-Id: I74779c47e8077d727a8a18c0e00696ba265f3c7a
Resolves: #78414
Releases: 8.4, master
Reviewed-on: https://review.typo3.org/50352
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frederic Gaus <frederic.gaus@flagbit.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>