[BUGFIX] Make extbase subclasses work with numeric types
The extbase subclasses configuration does currently not work with a
numeric database field that identifies the type of a record. This is
due to a strict comparison that compares the stringish TypoScript
value against the numeric database value.

Both values are casted to string to make the comparision work.

Resolves: #78416
Releases: master, 7.6
Change-Id: Iea4cf41301892a6788facd3befe5533e263cc3f6
Reviewed-on: https://review.typo3.org/50374
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>