[BUGFIX] Pass correct query parameter to image manipulation view
The TCA type "imageManipulation" has a boolean config called "enableZoom".
As declared in "ImageManipulationElement" this config will be added to
the "ImageManipulationWizard" wizardData in a key called "zoom".
But "ImageManipulationWizard" assigns a non existing key "bool".

So "enableZoom" set to true or false will always result in false.

Releases: master, 7.6
Resolves: #78444
Change-Id: I7580927d4edeec246d71ade32497d34ba107589d
Reviewed-on: https://review.typo3.org/50394
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>