[BUGFIX] Prevent installation of incompatible extensions
If an extension is installed, all information are gathered from database.
Here also a check for possible update versions is triggered. In this
check potential dependency updates are marked which get installed during
the underlying extension installation.
This patch prevents gathering TER information during an installation
process to prevent any update information to be registered.

Resolves: #78494
Resolves: #76542
Change-Id: I0fb2b4632bef05912a329553b70f26d614a1ba9c
Reviewed-on: https://review.typo3.org/50437
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>