[BUGFIX] Allow RTE transformations in palettes
When having an RTE inside a palette (e.g. sys_file_reference
description field changed to a RTE field) then the RTE configuration
is not taken into account, thus it's not possible to do proper
transformations. All links are not working then.

The patch allows to parse the fields inside palettes as well.

Resolves: #60399
Resolves: #76745
Releases: master, 7.6
Change-Id: I0c4efc0a03682cd766a2a164cf7b2fa14cd7c1b9
Reviewed-on: https://review.typo3.org/50465
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>