[BUGFIX] ElementBrowser does not highlight selected file in file list view
FileLinkHandler in element browser should highlight the selected file
reference in browser. If the link value is "file:12345" the part after
"file:" from the string will be extracted and has to be cast as an int.
The later check for CSS class "active" in file list compares two numbers,
so the ID must be of type integer.

Resolves: #78574
Releases: 7.6
Change-Id: I11f8bf6bc73a08ff72edcb4769a206010d0226af
Reviewed-on: https://review.typo3.org/50502
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>