[TASK] Refactor suggest wizard JS communication
Transmitting the full $row in suggest wizard was introduced to
find the relevant flex form data structure in the ajax search action
to determine the suggest wizard field configuration

The patch adds the data structure identifier instead, the searchAction()
uses that to fetch the specified flex form data structure directly.

Additionally, field values relevant for the ajax call are now signed with
hmac and checked in searchAction().

Change-Id: Ibb5004d4d11487fe999072999c7ce2f847aac750
Resolves: #78616
Related: #78581
Releases: master
Reviewed-on: https://review.typo3.org/50544
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>