[BUGFIX] Let Extension Manager API also create folders
The API method added in for #70606 does not
create extension folders, which makes the API
incomplete and more or less useless.

Move the method call to create the folder into the API method.
Although the creation is done later then, the code in question
didn't need the folders anyway.

Resolves: #78631
Related: #70606
Releases: master, 7.6
Change-Id: I17b3527cd2c26e0bb1bdfedcb68dc7ae6064500d
Reviewed-on: https://review.typo3.org/50632
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>