[BUGFIX] Do not serialize variables for TransientMemoryBackend
This change switches the internal behavior of the VariableFrontend
when combined with the TransientMemoryBackend. Before this
patch the VariableFrontend was only capable of storing strings
and would serialize all variables it received, and unserialize all
variables it retrieves.

Rather than do this unnecessary serializing, a new contract is
implemented to indicate that the backend implementing the
contract is capable of storing non-string values, including
references to objects. When the VariableFrontend detects
this contract it skips the serialize/unserialize steps.

Change-Id: I255dbda2ae3791ad6325c5b4ad67c97e172e22f4
Releases: master, 7.6
Resolves: #78664
Reviewed-on: https://review.typo3.org/50600
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>