[TASK] Remove iconv support and use mbstring-polyfill
We remove the iconv support in the CharsetConverter as the PHP internal
implementation of some functions differ from their mb_string pendants.
This gives us a real headache in edge-cases.

We already have the mbstring-polyfill as an indirect dependency in
the Core (by symfony), so we simply require it our own now and use
mb_string all over the places.

This renders quite a few methods of CharsetConverter useless
and we therefore deprecate those.

Resolves: #78670
Releases: master
Change-Id: I88479f7939e0afb46f704d4e8f347abd2e10fc6d
Reviewed-on: https://review.typo3.org/50607
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>