[BUGFIX] Bind the cHash to the id of the "real" page
If a link to a mount point or a shortcut is created,
the id argument in the url points to the id of the real page,
where the mount or the shortcut points to.

This is why we need to take the id of the final page
to be hashed and not the one used to generate the link.

We also need to move cHash calculation below shortcut resolving.

Resolves: #78238
Releases: master, 7.6, 6.2
Change-Id: I96c4305c6d769a5851c17d6fdcba3cd7584fc3ec
Reviewed-on: https://review.typo3.org/50624
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>