[TASK] Manual require_once of crawler removal
The crawler is still included via a require_once in indexed_search,
even though the class loader since TYPO3 6.2 does that transparently.

So, the method is still called, which makes it impossible to
move EXT:crawler lib to a different location inside the extension.

The patch removes the call to the method, as the autoloader
fetches the file anyway.

It also deprecates the method call within EXT:indexed_search.

Resolves: #78679
Releases: master
Change-Id: I9654b89a520fe0aefd55a713613db794b8c88f72
Reviewed-on: https://review.typo3.org/50643
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>