[BUGFIX] Add check for valid directory name for TCA select fileFolder
If an invalid extension name is passed to a 'EXT:...' path, the
GeneralUtility::getFileAbsFileName function returns an empty string.
This results in a scan of '\' in the first place, and the resulting
error is not helpful for finding the source of the missing icons
in the TCA select box.
A check for a valid directory name and an exception in case of
failure will help the situation.

Resolves: #78737
Releases: master, 7.6
Change-Id: I9d5159eac9b511e879144789eecb40eafcc0288c
Reviewed-on: https://review.typo3.org/50712
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>